refactor(ir): Use MakeTuple for shape and offset arguments
Replace variadic shape/offset arguments with MakeTuple expressions for
tensor and block operators. This provides better type safety and enables
more consistent tuple handling across the IR.

Changes:
- tensor.create, tensor.view, tensor.reshape, tensor.assemble
- block.view, block.reshape
- Update Python bindings to construct MakeTuple for shape/offset
- Add FormatShape helper for better error messages
- Fix PTO codegen to handle TupleGetItemExpr for constant extraction

+++ tests/ut/ir/operators/test_tensor_ops.py
@@ -392,13 +392,13 @@ def test_tensor_reshape_dynamic():
     span = ir.Span.unknown()
 
     # Create a tensor with dynamic dimensions
-    dim_n = ir.Var("n", ir.ScalarType(DataType.INT32), span)
-    dim_m = ir.Var("m", ir.ScalarType(DataType.INT32), span)
+    dim_n = ir.Var("n", ir.ScalarType(DataType.INT64), span)
+    dim_m = ir.Var("m", ir.ScalarType(DataType.INT64), span)
     tensor_type = ir.TensorType([dim_n, dim_m], DataType.FP16)
     tensor_var = ir.Var("t", tensor_type, span)
 
     # Reshape with dynamic shape (cannot verify element count at compile time)
-    dim_k = ir.Var("k", ir.ScalarType(DataType.INT32), span)
+    dim_k = ir.Var("k", ir.ScalarType(DataType.INT64), span)
     call = ir.op.tensor.reshape(tensor_var, [dim_k])
 
     assert isinstance(call, ir.Call)
